Add enter key to jump window
Enter key jumps to window in the first index of the program. This is done by
subclassing the edit box controls. Implementation is still exploratory and as
such the code is rough.

This build almost reaches feature parity with the original utility in ImGUI
+ GLFW3 with the exception of currently not filtering by individual word but
instead from the start of the string.